requirements.txt
pandas==2.0.1
numpy==1.24.3
openai==1.77.0
python-dotenv==1.0.0
scikit-learn==1.2.2
plotly==5.14.1
jupyterlab==4.0.0
matplotlib==3.7.1
tsne==0.1.3
